requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Firebase
firebase-admin==6.3.0

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
imbalanced-learn==0.11.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Authentication & Security
PyJWT==2.8.0
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
requests==2.31.0

# Deployment
gunicorn==21.2.0
waitress==2.1.2

# Development
pytest==7.4.3
